Refactor workspace logic into polymer-workspaces standalone library
