+++ package.json
@@ -30,47 +30,36 @@
   },
   "dependencies": {
     "@types/chai": "^3.5.2",
-    "@types/diff": "^3.2.0",
     "@types/esprima": "^2.1.33",
-    "@types/estraverse": "^0.0.6",
-    "@types/estree": "^0.0.35",
     "@types/inquirer": "^0.0.35",
-    "@types/mocha": "^2.2.41",
     "@types/mz": "^0.0.31",
-    "@types/progress": "^2.0.0",
-    "@types/rimraf": "0.0.28",
+    "@types/rimraf": "2.0.2",
     "@types/semver": "^5.3.32",
     "ast-types": "^0.9.11",
-    "bottleneck": "^1.16.0",
-    "bower": "^1.8.0",
     "chai": "^3.5.0",
+    "chalk": "^2.0.1",
     "command-line-args": "^4.0.6",
     "command-line-usage": "^4.0.0",
     "esprima": "^3.1.3",
-    "estraverse": "^4.2.0",
-    "github": "~9.2.0",
     "inquirer": "^3.2.0",
     "ix": "^2.0.1",
     "jscodeshift": "^0.3.30",
     "mkdirp": "^0.5.1",
-    "mocha": "^3.4.1",
     "mz": "^2.6.0",
-    "nodegit": "^0.18.3",
-    "pad": "^1.1.0",
     "polymer-analyzer": "^2.0.2",
-    "progress": "^2.0.0",
     "recast": "^0.12.4",
-    "resolve": "^1.3.3",
     "rimraf": "^2.6.1",
     "semver": "^5.3.0",
     "source-map-support": "^0.4.15",
     "tslib": "^1.7.1"
   },
   "devDependencies": {
-    "chalk": "^2.0.1",
+    "@types/diff": "^3.2.0",
+    "@types/mocha": "^2.2.41",
     "clang-format": "^1.0.53",
     "diff": "^3.3.0",
     "gulp": "^3.9.1",
+    "mocha": "^3.5.3",
     "tslint": "^5.5.0",
     "typescript": "^2.4.1",
     "watchy": "^0.7.0"
